webapp/requirements.txt
flask
tensorflow
opencv-python
numpy
pillow
